ultranest: allow MPI runs; add plotting; allow post-processing
Since ultranest already has MPI parallelisation to efficiently draw
samples, it is unnecessary for mosfit to also parallelise.
This uses serial mosfit processing in case ultranest is enabled.

In case files are stored into an output directory, this enables
the usual diagnostics and corner plots of the ultranest run.

Posterior samples are passed back to the fitter for plotting.